// webSyscsc/ClientApp/src/app/syscsc/models/agroclimatica.ts
export class Agroclimatica {
    agroclimaticaId: number;
    productorId: number;
    latitud: string;
    norteLongitud: string;
    este: string;
    msnm: string;
    analisisSuelo: string;
    fechaRealizacion: string;
    planFertilizacion: string;
    tipoTextura: string;
    preparacionAbono: string;
    tipo: string;
    estado: string;
    observaciones: string;
}